% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan-package.R
\docType{data}
\name{burndown}
\alias{burndown}
\title{Sample burndown dataset}
\format{See \code{\link{read.burndown}}.}
\description{
This is sample burndown dataset provided for testing.
}
\author{
Dan Kelley
}
\seealso{
\code{\link{read.burndown}} may be used to read such objects,
\code{\link{summary.burndown}} to summarize them, and
\code{\link{plot.burndown}} to plot them.
}
\keyword{datasets}

