% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{plan-class}
\alias{plan-class}
\title{Base Class for plan Objects}
\description{
Base Class for plan Objects
}
\section{Slots}{

\describe{
\item{\code{data}}{A list containing variable contents.}
}}

\concept{classes provided by plan}
