% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{[[,plan-method}
\alias{[[,plan-method}
\title{Extract Something From a plan Object}
\usage{
\S4method{[[}{plan}(x, i, j, ...)
}
\arguments{
\item{x}{A \linkS4class{plan} object.}

\item{i}{The item to extract.}

\item{j}{Optional additional information on the \code{i} item.}

\item{...}{Optional additional information (ignored).}
}
\description{
Extract something from a plan object, avoiding using the "slot" notation.
}
