% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{[[<-,plan-method}
\alias{[[<-,plan-method}
\title{Replace Parts of a plan Object}
\usage{
\S4method{[[}{plan}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A \linkS4class{plan} object.}

\item{i}{The item to replace.}

\item{j}{Optional additional information on the \code{i} item.}

\item{...}{Optional additional information (ignored).}

\item{value}{The value to be placed into \code{x}, somewhere.}
}
\description{
Replace something within a plan object, avoiding using the "slot" notation.
}
