% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{summary,gantt-method}
\alias{summary,gantt-method}
\title{Summarize a gantt object}
\usage{
\S4method{summary}{gantt}(object, ...)
}
\arguments{
\item{object}{A \linkS4class{gantt} object.}

\item{...}{ignored.}
}
\description{
Summarizes a gantt object.
}
\details{
Prints a summary of a gantt dataset.
}
\examples{
library(plan)
data(gantt)
summary(gantt)
}
\references{
\url{https://en.wikipedia.org/wiki/Burndown_chart}
}
\seealso{
Other things related to gantt data: 
\code{\link{as.gantt}()},
\code{\link{gantt}},
\code{\link{gantt-class}},
\code{\link{ganttAddTask}()},
\code{\link{plot,gantt-method}},
\code{\link{read.gantt}()}
}
\author{
Dan Kelley
}
\concept{things related to gantt data}
