% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burndown.R
\name{summary.burndown}
\alias{print.summary.burndown}
\alias{summary.burndown}
\title{Summarize a burndown object}
\usage{
\method{summary}{burndown}(object, ...)
}
\arguments{
\item{object}{an object of class \code{burndown}, e.g. as read by
\code{\link{read.burndown}}.}

\item{x}{an object of class \code{summary.burndown}, as created by
\code{summary.burndown}.}

\item{\dots}{extra arguments (not used in this version).}
}
\value{
None.
}
\description{
Summarizes a burndown object.
}
\details{
Prints a summary of a burndown dataset.
}
\examples{

library(plan)
data(burndown)
summary(burndown)

}
\author{
Dan Kelley
}
\seealso{
The \code{burndown} object may be read with
\code{\link{read.burndown}} and plotted with \code{\link{plot.burndown}}.
}
\keyword{misc}

