% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{summary.gantt}
\alias{print.summary.gantt}
\alias{summary.gantt}
\title{Summarize a gantt object}
\usage{
\method{summary}{gantt}(object, ...)
}
\arguments{
\item{object}{an object of class \code{gantt}, e.g. as read by
\code{\link{read.gantt}}.}

\item{x}{an object of class \code{summary.gantt}, as created by
\code{summary.gantt}.}

\item{\dots}{extra arguments (not used in this version).}
}
\value{
None.
}
\description{
Summarizes a gantt object.
}
\details{
Prints a summary of a gantt dataset.
}
\examples{

library(plan)
data(gantt)
summary(gantt)

}
\author{
Dan Kelley
}
\references{
\url{http://alistair.cockburn.us/crystal/articles/evabc/earnedvalueandburncharts.htm}.
}
\seealso{
The \code{gantt} object may be read with \code{\link{read.gantt}}.
}
\keyword{misc}

