% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burndown.R
\docType{class}
\name{burndown-class}
\alias{burndown-class}
\title{Class to store burndown objects}
\description{
Class to store burndown objects
}
\seealso{
Other things related to burndown data: 
\code{\link{as.burndown}()},
\code{\link{burndown}},
\code{\link{plot,burndown-method}},
\code{\link{read.burndown}()},
\code{\link{summary,burndown-method}}
}
\concept{things related to burndown data}
