% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan-package.R
\docType{data}
\name{burndown}
\alias{burndown}
\title{Sample burndown dataset}
\description{
This is sample burndown dataset provided for testing.
}
\seealso{
Other things related to \code{burndown} data: \code{\link{burndown-class}},
  \code{\link{plot,burndown-method}},
  \code{\link{read.burndown}},
  \code{\link{summary,burndown-method}}

Other data sets provided with \code{plan}.: \code{\link{gantt}}
}
\author{
Dan Kelley
}
