\name{burndown}

\docType{data}

\alias{burndown}

\title{Sample burndown dataset}

\description{This is sample burndown dataset provided for testing.}

\usage{data(burndown)}

\format{See \code{\link{read.burndown}}.}

\seealso{\code{\link{read.burndown}} may be used to read such objects,
  \code{\link{summary.burndown}} to summarize them, and
  \code{\link{plot.burndown}} to plot them.}

\author{Dan Kelley}

\keyword{datasets}
