% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\docType{class}
\name{gantt-class}
\alias{gantt-class}
\title{Class to store \code{gantt} objects}
\description{
These objects may be created with \code{\link{as.gantt}} or
\code{\link{read.gantt}}.
}
\seealso{
Other things related to \code{gantt} data: \code{\link{as.gantt}},
  \code{\link{ganttAddTask}}, \code{\link{gantt}},
  \code{\link{plot,gantt-method}},
  \code{\link{read.gantt}},
  \code{\link{summary,gantt-method}}
}
