% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan-package.R
\docType{data}
\name{gantt}
\alias{gantt}
\title{Sample gantt dataset}
\format{See \code{\link{read.gantt}}}
\description{
This is sample gantt dataset provided for testing.
}
\author{
Dan Kelley
}
\seealso{
\code{\link{read.gantt}} may be used to read such objects,
\code{\link{summary.gantt}} to summarize them, and \code{\link{plot.gantt}}
to plot them.
}
\keyword{datasets}

