% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan-package.R
\docType{data}
\name{gantt}
\alias{gantt}
\title{Sample gantt dataset}
\format{See \code{\link{read.gantt}}}
\description{
This is sample gantt dataset provided for testing.
}
\author{
Dan Kelley
}
\seealso{
Other data sets provided with \code{plan}.: \code{\link{burndown}}
}

