\name{gantt}

\docType{data}

\alias{gantt}

\title{Sample gantt dataset}

\description{This is sample gantt dataset provided for testing.}

\usage{data(gantt)}

\format{See \code{\link{read.gantt}}}

\seealso{\code{\link{read.gantt}} may be used to read such objects,
  \code{\link{summary.gantt}} to summarize them, and
  \code{\link{plot.gantt}} to plot them.}

\author{Dan Kelley}

\keyword{datasets}
