% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan-package.R
\docType{data}
\name{gantt}
\alias{gantt}
\title{Sample gantt dataset}
\description{
This is sample gantt dataset provided for testing.
}
\seealso{
Other things related to gantt data: 
\code{\link{as.gantt}()},
\code{\link{gantt-class}},
\code{\link{ganttAddTask}()},
\code{\link{plot,gantt-method}},
\code{\link{read.gantt}()},
\code{\link{summary,gantt-method}}

Other data sets provided with plan: 
\code{\link{burndown}}
}
\author{
Dan Kelley
}
\concept{data sets provided with plan}
\concept{things related to gantt data}
