% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan-package.R
\docType{package}
\name{plan}
\alias{plan}
\title{Plan, a package for project planning}
\description{
This package provides tools for project planning, e.g. burndown charts,
gantt diagrams, etc.; see \linkS4class{burndown} and
\linkS4class{gantt} for entries to the documentation.
}
\author{
Dan Kelley
}
