% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{[[,plan-method}
\alias{[[,plan-method}
\title{Extract Something From a plan Object}
\usage{
\S4method{[[}{plan}(x, i, j, ...)
}
\arguments{
\item{x}{A \code{plan} object, i.e. one inheriting from \code{\link{plan-class}}.}

\item{i}{The item to extract.}

\item{j}{Optional additional information on the \code{i} item.}

\item{...}{Optional additional information (ignored).}
}
\description{
Extract something from a plan object, avoiding using the "slot" notation.
}
