% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burndown.R
\docType{methods}
\name{summary,burndown-method}
\alias{summary,burndown-method}
\title{Summarize a burndown object}
\usage{
\S4method{summary}{burndown}(object, ...)
}
\arguments{
\item{object}{A \code{burndown} object, i.e. one inheriting from
\code{\link{burndown-class}}.}

\item{...}{ignored.}
}
\description{
Print a summary of a burndown dataset.
}
\examples{
library(plan)
data(burndown)
summary(burndown)
}
\seealso{
Other things related to \code{burndown} data: \code{\link{burndown-class}},
  \code{\link{burndown}},
  \code{\link{plot,burndown-method}},
  \code{\link{read.burndown}}
}
\author{
Dan Kelley
}
