% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\docType{methods}
\name{summary,gantt-method}
\alias{summary,gantt-method}
\title{Summarize a gantt object}
\usage{
\S4method{summary}{gantt}(object, ...)
}
\arguments{
\item{object}{A \code{gantt} object, i.e. one inheriting from \code{\link{gantt-class}}.}

\item{...}{ignored.}
}
\description{
Summarizes a gantt object.
}
\details{
Prints a summary of a gantt dataset.
}
\examples{
library(plan)
data(gantt)
summary(gantt)
}
\references{
\url{http://alistair.cockburn.us/crystal/articles/evabc/earnedvalueandburncharts.htm}.
}
\seealso{
Other things related to \code{gantt} data: \code{\link{as.gantt}},
  \code{\link{gantt-class}}, \code{\link{ganttAddTask}},
  \code{\link{gantt}}, \code{\link{plot,gantt-method}},
  \code{\link{read.gantt}}
}
\author{
Dan Kelley
}
